ml/flyte/requirements.txt
flytekit
pandas
opencv-python
numpy
kagglehub
google-cloud-storage
Pillow
matplotlib==3.5.3
seaborn
tensorflow~=2.16.0
keras
scikit-learn
kubernetes
flytekitplugins-pod